requirements.txt
# Excel处理
pandas==2.1.1
openpyxl==3.1.2

# Spotify API
spotipy==2.23.0

# 环境变量管理
python-dotenv==1.0.0

# 进度条显示
tqdm==4.66.1 